+++ indianapolis-model/src/main/java/com/gdn/recommendation/Vehicle.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Vehicle {
     private String name;
-    private double cbmCapacity;
-    private double minCbm;
+    private Float cbmCapacity;
+    private Float minCbm;
 }
